+++ packages/plugins/IconMenu/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## unpublished
 
 - Feature: Add new configuration parameter `component` to be able to interchange the UI component of this plugin.
+- Feature: Expose `IconMenuWrapper.vue` as a new export in order to use it in custom implementations.
+
 
 ## 1.3.1