+++ packages/plugins/IconMenu/CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## unpublished
+## 1.4.0
 
 - Feature: Add new configuration parameter `buttonComponent` to be able to interchange the UI component of the used (icon) buttons.
 - Feature: Expose `IconMenuButton.vue` as a new export in order to use it in custom implementations.